+++ src/index.js
@@ -6,4 +6,3 @@
  */
 export {default as AutosuggestUI} from "./lib/ui/autosuggest";
 export {default as AutosuggestStore} from "./lib/store/autosuggestStore";
-export {default as AutosuggestStaticStore} from "./lib/store/autosuggestStaticStore";