+++ src/lib/ui/autosuggest.jsx
@@ -2,7 +2,7 @@
  * @module st-react-autosuggest
  */
 import React from "react";
-import Autosuggest from "react-autosuggest";
+import Autosuggest from "st-react-typeahead-component";
 
 /**
  * This is the ui for the autosuggest. It requires a backing store as a property upon
@@ -11,30 +11,29 @@ import Autosuggest from "react-autosuggest";
  * @class AutosuggestUI
  */
 
-
-class AutosuggestComponent extends React.Component {
+class AutosuggestUI extends React.Component {
 
   constructor(props) {
     super(props);
     this.state = {
-      originalOnSuggestionSelected: (this.props.onSuggestionSelected) ? this.props.onSuggestionSelected : null,
-      wrappedProps: this.wrapProps(this.props),
-      lastValue: this.valueFromProps(this.props),
-      justSelected: false,
-      value: this.valueFromProps(this.props)
+      input: this.props.autosuggestStore.getDisplayValue(this.props.initialSuggestion),
+      options: [],
+      currentlySelected: null
     };
   }
 
   /**
    * Called when new properties are passed to this component
    * @param props new properties
-     */
+   */
   componentWillReceiveProps(nextProps) {
+    if (nextProps.initialSuggestion === this.props.initialSuggestion) {
+      return;
+    }
     this.setState({
-      wrappedProps: this.wrapProps(nextProps),
-      lastValue: this.valueFromProps(nextProps),
-      justSelected: false,
-      value: this.valueFromProps(nextProps)
+      input: nextProps.autosuggestStore.getDisplayValue(nextProps.initialSuggestion),
+      options: [],
+      currentlySelected: null
     });
   }
 
@@ -43,103 +42,89 @@ class AutosuggestComponent extends React.Component {
    * @method render
    */
   render() {
+    var OptionTemplate = this.props.autosuggestStore.getSuggestionTemplate();
     return (
-      <Autosuggest ref="suggest" {...this.state.wrappedProps} value={this.state.value} suggestions={this.props.autosuggestStore.getSuggestion.bind(this.props.autosuggestStore)} />
+      <div className="typeahead-container">
+        <Autosuggest
+          ref="typeahead"
+          inputValue={this.state.input}
+          options={this.state.options}
+          optionTemplate={OptionTemplate}
+          onChange={this.handleChange.bind(this)}
+          onOptionChange={this.handleOptionChange.bind(this)}
+          onOptionClick={this.handleOptionClick.bind(this)}
+          onFocus={this.onFocus.bind(this)}
+          onBlur={this.onBlur.bind(this)}
+          onDropdownClose={this.onDropdownClose.bind(this)}
+         />
+      </div>
+
     );
   }
 
-  /**
-   *
-   * @param props
-   * @returns String
-     */
-  valueFromProps(props) {
-    if (props.value) {
-      return props.value;
+  handleChange(event) {
+    var input = event.target.value;
+    this.setInputValue(input);
+    if (this.showWhen(input)) {
+      this.props.autosuggestStore.getSuggestion(input, (error, suggestions) => {
+        if (suggestions) {
+          this.setOptions(suggestions);
+        }
+      });
     }
-
-    if (props.inputAttributes && props.inputAttributes.value) {
-      return props.inputAttributes.value;
-    }
-
-    return "";
   }
 
   /**
-   * Wraps the props object method of intrests with local method
-   * @returns {{}}
-   */
-  wrapProps(propsToWrap) {
-    var wrappedProps = {};
-
-    for (var propName in propsToWrap) {
-      if (propsToWrap.hasOwnProperty(propName)) {
-        wrappedProps[propName] = propsToWrap[propName];
-      }
-    }
-
-    //wrapped the methods of interests
-    wrappedProps.showWhen = this.showWhen.bind(this);
-    wrappedProps.onSuggestionSelected = this.onSuggestionSelected.bind(this);
-    if (!wrappedProps.inputAttributes) {
-      wrappedProps.inputAttributes = {};
-    }
-
-    if (this.props.focusPlaceholder) {
-      wrappedProps.inputAttributes.onFocus = this.onFocus.bind(this);
-      wrappedProps.inputAttributes.onBlur = this.onBlur.bind(this);
-    }
-    delete wrappedProps.value;
-    return wrappedProps;
-  }
-
-  componentDidMount() {
-    this.refs.suggest.refs.input.focus();
-    this.refs.suggest.refs.input.blur();
-  }
-
-  /**
-   * Wrapper method to prevent suggestion from showing if the underlying value hasn't changed.
-   * This is added to show the auto suggest only if input text is changed.
-   * Underlying auto-suggest tries to suggest even when input is focused and that can lead to unnecessary n/w call if
-   * the input field is pre-populated.
-   * @see <a href=""></a>
+   * Should we attempt to lookup suggestions
    *
    * @param inputText text from the auto-suggest input
    * @return boolean true if suggestions should be shown else false
    */
   showWhen(inputText) {
-    if (this.props.suggestionsOnlyOnInputChange === true) {
-      if (inputText === this.state.lastValue) {
-        return false;
-      }
-    }
-
-    return this.props.showWhen(inputText);
+    return inputText.length > 2;
   }
 
-  onSuggestionSelected(selection, event) {
-    if (this.state.originalOnSuggestionSelected) {
-      this.state.originalOnSuggestionSelected(selection, event);
-    }
+  setInputValue(value) {
     this.setState({
-      lastValue: this.props.suggestionValue(selection),
-      justSelected: true
+      input: value
     });
+  }
+
+  setOptions(options) {
+    this.setState({
+      options: options
+    });
+  }
+
+  onDropdownClose() {
+    if (this.state.currentlySelected) {
+      this.props.onSelected(this.state.currentlySelected);
+      this.setState({
+        currentlySelected: null
+      });
+    }
+  }
+
+  handleOptionChange(event, option, index) {
+    if (index === -1) {
+      return;
+    }
+    this.setInputValue(this.props.autosuggestStore.getDisplayValue(option));
+    this.setState({
+      currentlySelected: option
+    });
+  }
+
+  handleOptionClick(event, option) {
+    this.setInputValue(this.props.autosuggestStore.getDisplayValue(option));
+    this.props.onSelected(option);
   }
 
   onFocus(evt) {
     var value = evt.target.value;
-
-    if (this.state.justSelected) {
-      this.setState({
-        justSelected: false
-      });
-      return;
-    }
-
     this.setState({
-      value: this.props.focusPlaceholderText,
+      input: this.props.focusPlaceholderText,
+      options: [],
       lastValueBeforeFocus: value
     });
   }
@@ -147,13 +132,14 @@ class AutosuggestComponent extends React.Component {
   onBlur(evt) {
     if (evt.target.value === this.props.focusPlaceholderText) {
       this.setState({
-        value: this.state.lastValueBeforeFocus
+        input: this.state.lastValueBeforeFocus,
+        options: []
       });
     }
   }
-};
+}
 
-AutosuggestComponent.propTypes = {
+AutosuggestUI.propTypes = {
   /**
    * @attribute autosuggestStore
    * @required
@@ -162,32 +148,33 @@ AutosuggestComponent.propTypes = {
   autosuggestStore: React.PropTypes.any.isRequired,
 
   /**
-   * @attribute focusPlaceholder
-   * @optional
-   * @type Boolean
+   * @attribute initialSuggestion
+   * @type Suggestion
    */
-  focusPlaceholder: React.PropTypes.bool,
+  initialSuggestion: React.PropTypes.any,
 
   /**
    * @attribute focusPlaceholderText
-   * @optional
    * @type String
+   * @default ""
    */
   focusPlaceholderText: React.PropTypes.string,
 
-
   /**
-   * @attribute  suggestionsOnlyOnInputChange
-   * @type Boolean
+   * @attribute onSelected
+   * @type Function
+   *   @param {Suggestion}
+   * @example
    */
-  suggestionsOnlyOnInputChange: React.PropTypes.bool
+  onSelected: React.PropTypes.func
+
 };
 
-AutosuggestComponent.defaultProps = {
+AutosuggestUI.defaultProps = {
   autosuggestStore: null,
-  focusPlaceholder: true,
+  initialSuggestion: null,
   focusPlaceholderText: "",
-  suggestionsOnlyOnInputChange: true
+  onSelected: (suggestion) => {}
 };
 
-export default AutosuggestComponent;
+export default AutosuggestUI;