+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+## 2.0.0 (2016-01-21)
+
+- Changed underlying autosuggest implementation to st-react-typeahead-component
+- Added keyboard support
+- Added input clear functionality on focus and revert on blur
+- Closing the selector will select the last focused option
+
 ## 1.1.0 (2016-01-19)
 
-- Added "suggestionsOnlyOnInputChange" property to autosuggest
-- Added input clear functionality on focus and revert on blur
+
 
 ## 1.0.0 (2015-12-01)